cleaned up inner css and scripts, changed index back to home and put product display in products.php.
