+++ front_end/css/products.css
@@ -1,0 +1,33 @@
+
+        body { font-family: Arial, sans-serif; margin: 20px; }
+        .product { 
+            border: 1px solid #ddd; 
+            padding: 15px; 
+            margin-bottom: 20px; 
+            border-radius: 5px;
+            display: flex;
+            gap: 20px;
+        }
+        .product img { 
+            max-width: 150px; 
+            max-height: 150px;
+            object-fit: contain;
+        }
+        .product-info {
+            flex: 1;
+        }
+        .price { 
+            color: green; 
+            font-weight: bold;
+            font-size: 1.2em;
+            margin: 10px 0;
+        }
+        .view-product {
+            display: inline-block;
+            padding: 8px 15px;
+            background: #0066cc;
+            color: white;
+            text-decoration: none;
+            border-radius: 4px;
+        }
+
